Some minor tidying up, mostly to make the code conform to two space indentation, and other minor style considerations.
--HG--
extra : convert_revision : svn%3Aeff31bef-be4a-0410-a8fe-e47997df2690/trunk%4013

+++ reverse.c
@@ -825,7 +825,7 @@ extern int write_packet_data(WRITER   output,
     {
       fprintf(stderr,"### Error writing out data: %s\n"
               "    Wrote %d bytes instead of %d\n",
-              strerror(errno),written,data_len);
+              strerror(errno),(int)written,data_len);
       return 1;
     }
   }
